Move map and bootprint code into BootprintMapFragment.
